Refactoring debugger into separate files and moving state to props
