+++ webapp/src/blocks.tsx
@@ -6,7 +6,6 @@ import * as pkg from "./package";
 import * as core from "./core";
 import * as toolboxeditor from "./toolboxeditor"
 import * as compiler from "./compiler"
-import * as debug from "./debugger";
 import * as toolbox from "./toolbox";
 import * as snippets from "./blocksSnippets";
 import * as workspace from "./workspace";
@@ -14,6 +13,7 @@ import * as simulator from "./simulator";
 import { CreateFunctionDialog, CreateFunctionDialogState } from "./createFunction";
 
 import Util = pxt.Util;
+import { DebuggerToolbox } from "./debuggerToolbox";
 
 export class Editor extends toolboxeditor.ToolboxEditor {
     editor: Blockly.WorkspaceSvg;
@@ -29,11 +29,10 @@ export class Editor extends toolboxeditor.ToolboxEditor {
     showCategories: boolean = true;
     breakpointsByBlock: pxt.Map<number>; // Map block id --> breakpoint ID
     breakpointsSet: number[]; // the IDs of the breakpoints set.
-    debuggerToolboxDiv: JSX.Element;
+
+    protected debuggerToolbox: DebuggerToolbox;
 
     public nsMap: pxt.Map<toolbox.BlockDefinition[]>;
-
-    private debugVariables: debug.DebuggerVariables;
 
     setBreakpointsMap(breakpoints: pxtc.Breakpoint[]): void {
         let map: pxt.Map<number> = {};
@@ -563,15 +562,17 @@ export class Editor extends toolboxeditor.ToolboxEditor {
         this.toolbox = c;
     }
 
+    handleDebuggerToolboxRef = (c: DebuggerToolbox) => {
+        this.debuggerToolbox = c;
+    }
+
     renderToolbox(immediate?: boolean) {
         if (pxt.shell.isReadOnly()) return;
         const blocklyToolboxDiv = this.getBlocklyToolboxDiv();
-        const debuggerToolboxDiv = <div id="debuggerToolbox"></div>;
         const blocklyToolbox = <div className="blocklyToolbox">
             <toolbox.Toolbox ref={this.handleToolboxRef} editorname="blocks" parent={this} />
-            {debuggerToolboxDiv}
+            {<div id="debuggerToolbox"></div>}
         </div>;
-        this.debuggerToolboxDiv = debuggerToolboxDiv;
         Util.assert(!!blocklyToolboxDiv);
         ReactDOM.render(blocklyToolbox, blocklyToolboxDiv);
 
@@ -579,20 +580,18 @@ export class Editor extends toolboxeditor.ToolboxEditor {
     }
 
     updateToolbox() {
-        if (!this.debuggerToolboxDiv) return; // nothing to do here
+        const container = document.getElementById('debuggerToolbox');
+        if (!container) return;
 
         const debugging = !!this.parent.state.debugging;
-        let debuggerToolbox = debugging ? <div>
-            <debug.DebuggerToolbar parent={this.parent} />
-            <debug.DebuggerVariables ref={this.handleDebuggerVariablesRef} parent={this.parent} apisByQName={this.blockInfo.apis.byQName} />
-        </div> : <div />;
-        Util.assert(!!this.debuggerToolboxDiv)
+        let debuggerToolbox = debugging ? <DebuggerToolbox ref={this.handleDebuggerToolboxRef} parent={this.parent} apis={this.blockInfo.apis.byQName} /> : <div />;
+
         if (debugging) {
             this.toolbox.hide();
         } else {
             this.toolbox.show();
         }
-        ReactDOM.render(debuggerToolbox, document.getElementById('debuggerToolbox'));
+        ReactDOM.render(debuggerToolbox, container);
     }
 
     showPackageDialog() {
@@ -803,28 +802,21 @@ export class Editor extends toolboxeditor.ToolboxEditor {
         return false;
     }
 
-    handleDebuggerVariablesRef = (c: debug.DebuggerVariables) => {
-        this.debugVariables = c;
-    }
-
     updateDebuggerVariables(brk: pxsim.DebuggerBreakpointMessage) {
         if (!this.parent.state.debugging) return;
 
-        if (this.debugVariables) {
+        if (this.debuggerToolbox) {
             const visibleVars = Blockly.Variables.allUsedVarModels(this.editor).map((variable: any) => variable.name as string);
 
-            if (brk) {
-                this.debugVariables.updateVariables(brk.globals, brk.stackframes, visibleVars)
-            }
-            else {
-                this.debugVariables.updateVariables(null, null, visibleVars)
-            }
+            this.debuggerToolbox.setState({
+                lastBreakpoint: brk,
+                varFilters: visibleVars
+            });
         }
     }
 
     clearHighlightedStatements() {
         this.editor.highlightBlock(null);
-        if (this.debugVariables) this.debugVariables.clear();
     }
 
     openTypeScript() {
