--- webapp/src/debugger.tsx
@@ -1,408 +1,0 @@
-import * as React from "react";
-import * as ReactDOM from 'react-dom';
-import * as sui from "./sui";
-import * as data from "./data";
-import * as simulator from "./simulator";
-
-const MAX_VARIABLE_LENGTH = 20;
-
-type ISettingsProps = pxt.editor.ISettingsProps;
-
-interface ScopeVariables {
-    title: string;
-    variables: Variable[];
-    key?: string;
-}
-
-interface DebuggerVariablesState {
-    globalFrame: ScopeVariables;
-    stackFrames: ScopeVariables[];
-
-    frozen?: boolean;
-}
-
-interface Variable {
-    name: string;
-    value: any;
-    prevValue?: any;
-    children?: Variable[];
-}
-
-interface DebuggerVariablesProps extends ISettingsProps {
-    apisByQName: pxt.Map<pxtc.SymbolInfo>;
-}
-
-export class DebuggerVariables extends data.Component<DebuggerVariablesProps, DebuggerVariablesState> {
-    constructor(props: DebuggerVariablesProps) {
-        super(props);
-        this.state = {
-            globalFrame: {
-                title: lf("Globals"),
-                variables: []
-            },
-            stackFrames: []
-        };
-
-        this.toggleDebugging = this.toggleDebugging.bind(this);
-    }
-
-    clear() {
-        this.setState({
-            globalFrame: {
-                title: this.state.globalFrame.title,
-                variables: []
-            },
-            stackFrames: []
-        });
-    }
-
-    update(frozen = false) {
-        this.setState({ frozen });
-    }
-
-    toggleDebugging(): void {
-        this.props.parent.toggleDebugging();
-    }
-
-    renderCore() {
-        const { globalFrame, stackFrames, frozen } = this.state;
-        const variableTableHeader = lf("Variables");
-        let variables = globalFrame.variables;
-
-        // Add in the local variables.
-        // TODO: Handle callstack
-        if (stackFrames && stackFrames.length) {
-            variables = stackFrames[0].variables.concat(variables);
-        }
-
-        return <div className={`ui varExplorer`}>
-            <div className="ui variableTableHeader">
-                {variableTableHeader}
-            </div>
-            <div className={`ui segment debugvariables ${frozen ? "frozen" : ""} ui collapsing basic striped table`}>
-                {variables.length ? this.renderVariables(variables) : <div />}
-            </div>
-        </div>;
-    }
-
-    updateVariables(globals: pxsim.Variables, stackFrames: pxsim.StackFrameInfo[], filters?: string[]) {
-        if (!globals) {
-            // freeze the ui
-            this.update(true)
-            return;
-        }
-
-        const updatedGlobals = updateScope(this.state.globalFrame, globals);
-        if (filters) {
-            updatedGlobals.variables = updatedGlobals.variables.filter(v => filters.indexOf(v.name) !== -1)
-        }
-
-        let updatedFrames: ScopeVariables[];
-        if (stackFrames) {
-            const oldFrames = this.state.stackFrames;
-
-            updatedFrames = stackFrames.map(sf => {
-                const key = getKeyForFrame(sf);
-
-                for (const frame of oldFrames) {
-                    if (frame.key === key) return updateScope(frame, sf.locals, getArgArray(sf.arguments));
-                }
-
-                return updateScope({ key, title: sf.funcInfo.functionName, variables: [] }, sf.locals, getArgArray(sf.arguments))
-            });
-        }
-
-        this.setState({
-            globalFrame: updatedGlobals,
-            stackFrames: updatedFrames || [],
-            frozen: false
-        });
-
-        function getArgArray(info: pxsim.FunctionArgumentsInfo): Variable[] {
-            if (info) {
-                return [{ name: "this", value: info.thisParam }, ...info.params]
-            }
-            return []
-        }
-    }
-
-    private renderVariables(variables: Variable[], parent?: string, depth?: number): JSX.Element[] {
-        let r: JSX.Element[] = [];
-        if (!parent) {
-            variables = variables.sort((a, b) => {
-                return variableType(a).localeCompare(variableType(b)) || a.name.toLowerCase().localeCompare(b.name.toLowerCase());
-            });
-        }
-        depth = depth || 0;
-        let margin = depth * 0.75 + 'em';
-        variables.forEach(v => {
-            const newValue = renderValue(v.value);
-            let type = variableType(v);
-            const onClick = v.value && v.value.id ? () => this.toggle(v) : undefined;
-
-            r.push(<div key={(parent || "") + v.name} role="listitem" className="item" onClick={onClick}>
-                <div className="variableAndValue">
-                    <div className={`variable varname ${v.prevValue !== undefined ? "changed" : ""}`} title={v.name}>
-                        <i className={`${(v.children ? "small down triangle icon" : "small right triangle icon") + ((v.value && v.value.hasFields) ? "" : " transparent")}`} style={{ marginLeft: margin }} ></i>
-                        <span>{v.name + ':'}</span>
-                    </div>
-                    <div className="variable detail" style={{ padding: 0.2 }} title={shouldShowValueOnHover(type) ? newValue : ""}>
-                        <span className={`varval ${type}`}>{truncateLength(newValue)}</span>
-                    </div>
-                </div>
-            </div>
-            );
-            if (v.children)
-                r = r.concat(this.renderVariables(v.children, v.name, depth + 1));
-        })
-        return r;
-    }
-
-    private toggle(v: Variable) {
-        // We have to take care of the logic for nested looped variables. Currently they break this implementation.
-        if (v.children) {
-            delete v.children;
-            this.setState({ globalFrame: this.state.globalFrame })
-        } else {
-            if (!v.value.id) return;
-            // We filter the getters we want to call for this variable.
-            let allApis = this.props.apisByQName;
-            let matcher = new RegExp("^((.+\.)?" + v.value.type + ")\.");
-            let potentialKeys = Object.keys(allApis).filter(key => matcher.test(key));
-            let fieldsToGet: string[] = [];
-
-            potentialKeys.forEach(key => {
-                let commentAttrs = allApis[key];
-                if (!key.endsWith("@set") && commentAttrs && commentAttrs.attributes.callInDebugger) {
-                    fieldsToGet.push(key);
-                }
-            });
-
-            simulator.driver.variablesAsync(v.value.id, fieldsToGet)
-                .then((msg: pxsim.VariablesMessage) => {
-                    if (msg && msg.variables) {
-                        v.children = Object.keys(msg.variables).map(key => ({ name: key, value: msg.variables[key] }))
-                        this.setState({ globalFrame: this.state.globalFrame })
-                    }
-                })
-        }
-    }
-}
-
-export interface DebuggerToolbarProps extends ISettingsProps {
-}
-
-export interface DebuggerToolbarState {
-    isDragging?: boolean;
-    xPos?: number;
-}
-
-export class DebuggerToolbar extends data.Component<DebuggerToolbarProps, DebuggerToolbarState> {
-
-    constructor(props: DebuggerToolbarProps) {
-        super(props);
-        this.state = {
-        }
-
-        this.restartSimulator = this.restartSimulator.bind(this);
-        this.dbgPauseResume = this.dbgPauseResume.bind(this);
-        this.dbgStepOver = this.dbgStepOver.bind(this);
-        this.dbgStepInto = this.dbgStepInto.bind(this);
-        this.dbgStepOut = this.dbgStepOut.bind(this);
-        this.exitDebugging = this.exitDebugging.bind(this);
-    }
-
-    restartSimulator() {
-        pxt.tickEvent('debugger.restart', undefined, { interactiveConsent: true });
-        this.props.parent.restartSimulator();
-    }
-
-    exitDebugging() {
-        pxt.tickEvent('debugger.exit', undefined, { interactiveConsent: true });
-        this.props.parent.toggleDebugging();
-    }
-
-    dbgPauseResume() {
-        pxt.tickEvent('debugger.pauseresume', undefined, { interactiveConsent: true });
-        this.props.parent.dbgPauseResume();
-    }
-
-    dbgStepOver() {
-        pxt.tickEvent('debugger.stepover', undefined, { interactiveConsent: true });
-        this.props.parent.dbgStepOver();
-    }
-
-    dbgStepInto() {
-        pxt.tickEvent('debugger.stepinto', undefined, { interactiveConsent: true });
-        this.props.parent.dbgStepInto();
-    }
-
-    dbgStepOut() {
-        pxt.tickEvent('debugger.stepout', undefined, { interactiveConsent: true });
-        simulator.dbgStepOut();
-    }
-
-    getMenuDom() {
-        const node = ReactDOM.findDOMNode(this);
-        return node && node.firstElementChild;
-    }
-
-    renderCore() {
-        const parentState = this.props.parent.state;
-
-        const simState = parentState.simState;
-        const isRunning = simState == pxt.editor.SimState.Running;
-        const isStarting = simState == pxt.editor.SimState.Starting;
-        const isDebugging = parentState.debugging;
-        if (!isDebugging) return <div />;
-
-        const isDebuggerRunning = simulator.driver && simulator.driver.state == pxsim.SimulatorState.Running;
-        const advancedDebugging = this.props.parent.isJavaScriptActive();
-
-        const isValidDebugFile = advancedDebugging || this.props.parent.isBlocksActive();
-        if (!isValidDebugFile) return <div />;
-
-        const dbgStepDisabled = isDebuggerRunning || isStarting;
-        const dbgStepDisabledClass = dbgStepDisabled ? "disabled" : ""
-
-        const restartTooltip = lf("Restart debugging");
-        const dbgPauseResumeTooltip = isRunning ? lf("Pause execution") : lf("Continue execution");
-        const dbgStepIntoTooltip = lf("Step into");
-        const dbgStepOverTooltip = lf("Step over");
-        const dbgStepOutTooltip = lf("Step out");
-
-        if (!isDebugging) {
-            return <div className="debugtoolbar" role="complementary" aria-label={lf("Debugger toolbar")} />
-        } else if (advancedDebugging) {
-            // Debugger Toolbar for the monaco editor.
-            return <div className="debugtoolbar" role="complementary" aria-label={lf("Debugger toolbar")}>
-                {!isDebugging ? undefined :
-                    <div className={`ui compact borderless menu icon`}>
-                        <sui.Item key='dbgpauseresume' className={`dbg-btn dbg-pause-resume ${dbgStepDisabledClass} ${isDebuggerRunning ? "pause" : "play"}`} icon={`${isDebuggerRunning ? "pause blue" : "play green"}`} title={dbgPauseResumeTooltip} onClick={this.dbgPauseResume} />
-                        <sui.Item key='dbgstepover' className={`dbg-btn dbg-step-over ${dbgStepDisabledClass}`} icon={`xicon stepover ${isDebuggerRunning ? "disabled" : "blue"}`} title={dbgStepOverTooltip} onClick={this.dbgStepOver} />
-                        <sui.Item key='dbgstepinto' className={`dbg-btn dbg-step-into ${dbgStepDisabledClass}`} icon={`xicon stepinto ${isDebuggerRunning ? "disabled" : ""}`} title={dbgStepIntoTooltip} onClick={this.dbgStepInto} />
-                        <sui.Item key='dbgstepout' className={`dbg-btn dbg-step-out ${dbgStepDisabledClass}`} icon={`xicon stepout ${isDebuggerRunning ? "disabled" : ""}`} title={dbgStepOutTooltip} onClick={this.dbgStepOut} />
-                        <sui.Item key='dbgrestart' className={`dbg-btn dbg-restart right`} icon={`refresh green`} title={restartTooltip} onClick={this.restartSimulator} />
-                    </div>}
-            </div>;
-        } else {
-            // Debugger Toolbar for the blocks editor.
-            return <div className="debugtoolbar" role="complementary" aria-label={lf("Debugger toolbar")}>
-                <div className={`ui compact borderless menu icon`}>
-                    <sui.Item key='dbgstep' className={`dbg-btn dbg-step separator-after ${dbgStepDisabledClass}`} icon={`arrow right ${dbgStepDisabled ? "disabled" : "blue"}`} title={dbgStepIntoTooltip} onClick={this.dbgStepInto} text={"Step"} />
-                    <sui.Item key='dbgpauseresume' className={`dbg-btn dbg-pause-resume ${isDebuggerRunning ? "pause" : "play"}`} icon={`${isDebuggerRunning ? "pause blue" : "play green"}`} title={dbgPauseResumeTooltip} onClick={this.dbgPauseResume} />
-                    <sui.Item key='dbgrestart' className={`dbg-btn dbg-restart`} icon={`refresh green`} title={restartTooltip} onClick={this.restartSimulator} />
-                </div>
-            </div>;
-        }
-    }
-}
-
-function updateScope(lastScope: ScopeVariables, newVars: pxsim.Variables, params?: Variable[]): ScopeVariables {
-    let current = Object.keys(newVars).map(varName => ({ name: fixVarName(varName), value: newVars[varName] }));
-
-    if (params) {
-        current = params.concat(current);
-    }
-
-    return {
-        ...lastScope,
-        variables: getUpdatedVariables(lastScope.variables, current)
-    };
-}
-
-function fixVarName(name: string) {
-    return name.replace(/___\d+$/, "");
-}
-
-function getUpdatedVariables(previous: Variable[], current: Variable[]): Variable[] {
-    return current.map(v => {
-        const prev = getVariable(previous, v);
-        if (prev && prev.value && !prev.value.id && prev.value !== v.value) {
-            return {
-                ...v,
-                prevValue: prev.value
-            }
-        }
-        return v
-    });
-};
-
-function getVariable(variables: Variable[], value: Variable) {
-    for (let i = 0; i < variables.length; i++) {
-        if (variables[i].name === value.name) {
-            return variables[i];
-        }
-    }
-    return undefined;
-}
-
-function renderValue(v: any): string {
-    let sv = '';
-    let type = typeof v;
-    switch (type) {
-        case "undefined": sv = "undefined"; break;
-        case "number": sv = v + ""; break;
-        case "boolean": sv = v + ""; break;
-        case "string": sv = JSON.stringify(v); break;
-        case "object":
-            if (v == null) sv = "null";
-            else if (v.text) sv = v.text;
-            else if (v.id && v.preview) return v.preview;
-            else if (v.id !== undefined) sv = "(object)"
-            else sv = "(unknown)"
-            break;
-    }
-    return sv;
-}
-
-function truncateLength(varstr: string) {
-    let remaining = MAX_VARIABLE_LENGTH - 3; // acount for ...
-    let hasQuotes = false;
-    if (varstr.indexOf('"') == 0) {
-        remaining -= 2;
-        hasQuotes = true;
-        varstr = varstr.substring(1, varstr.length - 1);
-    }
-    if (varstr.length > remaining)
-        varstr = varstr.substring(0, remaining) + '...';
-    if (hasQuotes) {
-        varstr = '"' + varstr + '"'
-    }
-    return varstr;
-}
-
-function variableType(variable: Variable): string {
-    let val = variable.value;
-    if (val == null) return "undefined";
-    let type = typeof val
-    switch (type) {
-        case "string":
-        case "number":
-        case "boolean":
-            return type;
-        case "object":
-            if (val.type) return val.type;
-            if (val.preview) return val.preview;
-            if (val.text) return val.text;
-            return "object";
-        default:
-            return "unknown";
-    }
-}
-
-function shouldShowValueOnHover(type: string): boolean {
-    switch (type) {
-        case "string":
-        case "number":
-        case "boolean":
-        case "array":
-            return true;
-        default:
-            return false;
-    }
-}
-
-function getKeyForFrame(sf: pxsim.StackFrameInfo) {
-    return sf.funcInfo.functionName + ";" + sf.funcInfo.fileName + ";" + sf.funcInfo.start;
-}
